+++ core/src/library/mod.rs
@@ -1,4 +1,5 @@
 // pub(crate) mod cat;
+mod actors;
 mod config;
 #[allow(clippy::module_inception)]
 mod library;
@@ -6,6 +7,7 @@ mod manager;
 mod name;
 
 // pub use cat::*;
+pub use actors::*;
 pub use config::*;
 pub use library::*;
 pub use manager::*;
